Move MissingServiceRuntimeException to a top level class
Fixes #28
